Outsource get, set and unset functions
